requirements.txt
keras
tensorflow
numpy
opencv-python
os
pandas
Pillow
scikit-learn
Flask
Werkzeug
json
pprint
bson
Flask-PyMongo